+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+
+## [1.0.0-beta6] - 2025-08-01
 ### Added
 - MultitronFactory
 
@@ -14,7 +18,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CommandDeps as the TaskCommand's only dependency
 - individual framework extensions/bridges
 - Process exit code handling
-
 
 ## [1.0.0-beta5] - 2025-06-24
 ### Added
@@ -61,7 +64,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2024-08-20
 - After months of development, the first release of Multitron is here!
 
-[Unreleased]: https://github.com/riki137/multitron/compare/1.0.0-beta5...master
+[Unreleased]: https://github.com/riki137/multitron/compare/1.0.0-beta6...master
+[1.0.0-beta6]: https://github.com/riki137/multitron/compare/1.0.0-beta5...1.0.0-beta6
 [1.0.0-beta5]: https://github.com/riki137/multitron/compare/1.0.0-beta4...1.0.0-beta5
 [1.0.0-beta4]: https://github.com/riki137/multitron/compare/1.0.0-beta3...1.0.0-beta4
 [1.0.0-beta3]: https://github.com/riki137/multitron/compare/1.0.0-beta2...1.0.0-beta3